\name{ant-package}
\Rdversion{1.1}
\alias{ant-package}
\docType{package}
\title{
R specific version of ant
}
\description{
Version of the apache ant build tool tweaked with a few R specific tasks
}
\details{
\tabular{ll}{
Package: \tab ant\cr
Type: \tab Package\cr
Version: \tab 0.0-5\cr
Date: \tab 2009-09-07\cr
License: \tab Apache (>=2.0)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Romain Francois <francoisromain@free.fr>

Maintainer: Romain Francois <francoisromain@free.fr>
}
\references{
	The Apache ant project: \url{http://ant.apache.org/}
	
	\link[ant:../javadoc/index]{The Java API of this package}
}
\keyword{ package }

